Fix circular imports - extract method to ntlinks_utils.py
